✨ Added support for default values
In order to comply with a recent request, this commit now adds the
possibility to use default values in case the desired environment
variable is not set.
As default values in my opinion are not exactly good practice for
environment variables that should be set, a warning will be shown every
time the default value needs to be used.